backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
python-jose[cryptography]==3.4.0
bcrypt==4.1.3
python-dotenv==1.1.0
python-multipart==0.0.9
pydantic[email]==2.11.3
qrcode[pil]==8.2
requests==2.32.3
beautifulsoup4==4.14.3
lxml==6.0.2
itsdangerous==2.2.0
google-api-python-client==2.122.0
google-auth==2.49.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.3.0
aiofiles==25.1.0
pywebpush==2.0.1
